refactor: move handleClose to search results
