+++ src/widgets/search-results/ui/SearchResults.module.scss
@@ -1,7 +1,7 @@
 .container {
   display: flex;
   align-items: flex-start;
-  gap: var(--space-m);
+  gap: var(--space-xl);
 }
 
 .list {
