+++ src/widgets/search-results/ui/SearchResults.tsx
@@ -1,4 +1,11 @@
-import { useNavigate, useSearchParams, Outlet } from 'react-router';
+import type { MouseEvent } from 'react';
+import {
+  useNavigate,
+  useSearchParams,
+  Outlet,
+  useLocation,
+  useParams,
+} from 'react-router';
 import { useCharacters, QueryParams } from '@/shared/api/characters';
 import { CharacterCardList } from '@/entities/character';
 import { Pagination } from '@/features/pagination';
@@ -6,7 +13,10 @@ import styles from './SearchResults.module.scss';
 
 export const SearchResults = () => {
   const [searchParams] = useSearchParams();
+  const { id } = useParams<{ id: string }>();
+  const location = useLocation();
   const navigate = useNavigate();
+
   const query = searchParams.get(QueryParams.Name) || '';
   const page = Number(searchParams.get(QueryParams.Page)) || 1;
 
@@ -21,6 +31,22 @@ export const SearchResults = () => {
     });
   };
 
+  const handleClose = (event: MouseEvent<HTMLDivElement>) => {
+    const target = event.target as HTMLElement;
+
+    if (id) {
+      if (
+        target === event.currentTarget ||
+        target.classList.contains(styles.list)
+      ) {
+        navigate({
+          pathname: '/',
+          search: location.search,
+        });
+      }
+    }
+  };
+
   return (
     <>
       <Pagination
@@ -30,7 +56,7 @@ export const SearchResults = () => {
         disabled={isLoading}
         className={styles.pagination}
       />
-      <div className={styles.container}>
+      <div className={styles.container} onClick={handleClose}>
         <CharacterCardList
           characters={data?.results || []}
           isLoading={isLoading}
